+++ packages/reload-middleware/src/ui5/middleware.ts
@@ -8,7 +8,7 @@ import type { ReloaderConfig, HttpsOptions } from '../base';
 import type { MiddlewareParameters } from '@ui5/server';
 import type { RequestHandler } from 'express';
 import { ToolsLogger, UI5ToolingTransport } from '@sap-ux/logger';
-import { resolve } from 'path';
+import { resolve } from 'node:path';
 import { watchManifestChanges } from '../base/livereload';
 
 module.exports = async ({ options, middlewareUtil }: MiddlewareParameters<ReloaderConfig>): Promise<RequestHandler> => {