+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.43
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.39.5
+
+## 0.1.42
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+-   Updated dependencies [43a2446]
+    -   @sap-ux/fe-fpm-writer@0.39.4
+
 ## 0.1.41
 
 ### Patch Changes