+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.15.14
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.48
+
+## 0.15.13
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+-   Updated dependencies [43a2446]
+    -   @sap-ux/project-input-validator@0.6.24
+    -   @sap-ux/inquirer-common@0.7.47
+    -   @sap-ux/project-access@1.32.3
+    -   @sap-ux/ui5-info@0.12.4
+
 ## 0.15.12
 
 ### Patch Changes