+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.5.3
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.48
+
+## 0.5.2
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+-   Updated dependencies [43a2446]
+    -   @sap-ux/project-input-validator@0.6.24
+    -   @sap-ux/inquirer-common@0.7.47
+    -   @sap-ux/ui5-info@0.12.4
+
 ## 0.5.1
 
 ### Patch Changes